클라이언트 최신으로 업데이트 완료

+++ LegendOfDoodles/LegendOfDoodles/Code/05.Objects/97.Skeleton/Skeleton.h
@@ -12,7 +12,7 @@ struct Bone {
 	int m_iParentID;
 	XMFLOAT4X4  m_m4x4Matrix;
 	std::vector<KeyFrame> m_Keyframe;
-	
+
 	XMFLOAT4X4 GetFrame(int frame) {
 		if (m_Keyframe.size() == 0) {
 			return Matrix4x4::Identity();
@@ -30,8 +30,8 @@ public: // 생성자, 소멸자
 
 public: // 공개 함수
 	int GetAnimationLength() { return m_iAnimationLegth; }
-	Bone GetBone(float id) { return m_vBone.at(id); }
-	int GetBoneCount() { return m_vBone.size(); }
+	Bone GetBone(float id) { return m_vBone.at(static_cast<size_t>(id)); }
+	int GetBoneCount() { return static_cast<int>(m_vBone.size()); }
 
 private:	 // 변수
 	int m_iAnimationLegth;
